requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.5
django-filter==22.1
django-redis==5.2.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.14
zope.interface==5.5.2
